requirements.txt
streamlit==1.31.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
xgboost==1.7.6
lightgbm==3.3.5
joblib==1.2.0
pytz==2023.3
supabase-py==2.0.0
